Move prompts to UI
